requirements.txt
Flask==2.0.2
mysql_connector_repackaged==0.3.1
python-dotenv==0.20.0
